pyproject.toml
[tool.setuptools]
py-modules = ["agentwatch"]

[tool.poetry]
name = "agentwatch"
version = "1.0.0"
description = "Platform agnostic Agentic AI runtime observability framework"
authors = ["Shai Dvash <shai.dvash@cyberark.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.11"
asyncio = "^3.4.3"
pydantic = "^2.10.6"
aiofiles = "^24.1.0"
aiohttp = "^3.11.14"
jsonpath-ng = "^1.7.0"
httpx = "^0.28.1"
uvicorn = "^0.34.0"
dotenv = "^0.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
autoflake = "^2.3.1"
mypy = "^1.15.0"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
langgraph = "^0.3.8"
langchain-anthropic = "^0.3.9"
websockets = "^15.0.1"
fastapi = "^0.115.11"
autogen-agentchat = ">=0.2,<1.0"
langchain-openai = "^0.3.9"
langchain-ollama = "^0.3.0"

[tool.pytest.ini_options]
pythonpath = [
  "src/"
]

[tool.poetry.scripts]
agentwatch = "agentwatch.cli:main"

[build-system]
requires = ["setuptools>=40.9.0", "wheel", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"